requirements.txt
gradio
langchain-community
langchain-openai
langchain-chroma>=0.1.2
pypdf
guardrails-ai>=0.5.13
langchain-text-splitters